cjet: don't check on set set whether we (our process) are the owner of the state.
Owner might use change instead of set. Which would be much more efficient.
But there are scenarios that do need this strange handling. For example several tasks are running in one eventloop.
No task knows about the others. One tasks is owner of a state and one wants to set the same state. Both do exist in the same process.