Don't cache request values when setting a state.
Previously, the owner peer of a state sent a change notification only
when the the owner peer adapted the requested value. Therefore, the
requested value had to be cached in the routing request inside the jet
daemon.

To reduce the overall complexity, it jet protocol now requires that an
owner peer always has to send a change notification. Thus, caching of
the requested value in the jet daemon is no longer required.