Support timeouts for routed requests when adding a state/method.
Deamon implementations have to implement a timeout for routed request
(set/call). Otherwise, if a method/state owner does not respond to a
routed request, the routed request itself might linger forever in the
jet daemon.

This extension allows a method/state owner to request a different
timeout for this specific method/state to be added.