Support timeouts for routed requests when setting a state or calling a method.
Deamon implementations have to implement a timeout for routed request
(set/call). Otherwise, if a method/state owner does not respond to a
routed request, the routed request itself might linger forever in the
jet daemon.

This extension allows set/call to request a different timeout for this
specific routed request.